requirements.txt
certifi==2026.1.4
charset-normalizer==3.4.4
idna==3.11
jaraco.classes==3.4.0
jaraco.context==6.1.0
jaraco.functools==4.4.0
keyring==25.7.0
more-itertools==10.8.0
python-slugify==8.0.4
requests==2.32.5
text-unidecode==1.3
unicat==0.7.14
urllib3==2.6.3
